+++ src/store/error.rs
@@ -19,7 +19,7 @@
 //! Error for the store.
 
 /// Dynamic error type of an [`super::PropertyStore`].
-type DynError = Box<dyn std::error::Error>;
+type DynError = Box<dyn std::error::Error + Send + Sync + 'static>;
 
 /// Error that wraps the type returned by an implementation of the [`super::PropertyStore`] trait.
 #[non_exhaustive]